Refactor: 패키지 구분
- 각 클래스별로 패키지를 구분함.
- Controller 클래스는 전체적인 application의 흐름을 관리하는 기능을 가지기에 이름은 ChristmasPromotionApplication으로 변경함.
