+++ src/main/java/christmas/controller/request/RequestOrdersController.java
@@ -1,8 +1,8 @@
-package christmas.controller;
+package christmas.controller.request;
 
-import christmas.Mapper;
-import christmas.Orders;
+import christmas.domain.Orders;
 import christmas.dto.OrdersDTO;
+import christmas.util.Mapper;
 import christmas.view.InputView;
 
 public class RequestOrdersController extends RequestController<Orders> {
