+++ src/main/java/christmas/domain/Badge.java
@@ -1,4 +1,4 @@
-package christmas;
+package christmas.domain;
 
 import java.util.Arrays;
