+++ src/main/java/christmas/domain/VisitedDate.java
@@ -1,4 +1,4 @@
-package christmas;
+package christmas.domain;
 
 import christmas.exception.VisitedDateException;
