+++ src/main/java/christmas/dto/MenuAndAmount.java
@@ -1,4 +1,4 @@
-package christmas;
+package christmas.dto;
 
 public record MenuAndAmount(String menu, String amount) {
 }
