+++ src/main/java/christmas/dto/OrdersDTO.java
@@ -1,6 +1,5 @@
 package christmas.dto;
 
-import christmas.MenuAndAmount;
 import java.util.List;
 
 public record OrdersDTO(List<MenuAndAmount> requestOrders) {
