+++ src/main/java/christmas/util/Validator.java
@@ -1,4 +1,4 @@
-package christmas;
+package christmas.util;
 
 import java.util.regex.Pattern;
