+++ src/test/java/christmas/util/GiveawayEventTest.java
@@ -1,7 +1,8 @@
-package christmas;
+package christmas.util;
 
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 
+import christmas.util.GiveawayEvent;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
