+++ src/test/java/christmas/util/MenusValidatorTest.java
@@ -1,8 +1,9 @@
-package christmas;
+package christmas.util;
 
 import static org.assertj.core.api.AssertionsForClassTypes.assertThatThrownBy;
 import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
 
+import christmas.dto.MenuAndAmount;
 import java.util.List;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
